Restructure codebase as module
