+++ src/ebookreader/go.mod
@@ -1,0 +1,3 @@
+module ebookreader
+
+go 1.24.2
